+++ Boletin 9.5/src/boletin/pkg9/pkg5/Pares.java
@@ -2,7 +2,7 @@ package boletin.pkg9.pkg5;
 
 public class Pares {
     
-    public void sucesionPares(int rep){
+    public void sucesionPares(double rep){
         int par= 2;
         int num = 1;
         System.out.println("Sucesio de pares:");
